+++ patterns/footer-navigation.php
@@ -8,8 +8,8 @@
 <!-- wp:group {"style":{"spacing":{"margin":{"top":"var:preset|spacing|70"},"padding":{"bottom":"var:preset|spacing|40"}}},"className":"site-component-footer has-global-padding is-layout-constrained"} -->
 <div class="wp-block-group site-component-footer has-global-padding is-layout-constrained" style="margin-top:var(--wp--preset--spacing--70);padding-bottom:var(--wp--preset--spacing--40)">
 
-    <!-- wp:paragraph {"align":"center"} -->
-    <p class="has-text-align-center" style="font-size:1rem"><a href="/impressum"><?php echo __( 'Imprint', 'mdb-theme-fse' ); ?></a></p>
+    <!-- wp:paragraph {"align":"center","style":{"typography":{"fontSize":"14px","fontWeight":"600","textTransform":"uppercase"}}} -->
+    <p class="has-text-align-center" style="font-size:14px;font-weight:600;text-transform:uppercase"><a href="/impressum"><?php echo __( 'Imprint', 'mdb-theme-fse' ); ?></a></p>
     <!-- /wp:paragraph -->
 
     <!-- wp:buttons {"layout":{"type":"flex","flexWrap":"nowrap","justifyContent":"center"},"style":{"spacing":{"margin":{"bottom":"0"},"blockGap":"2rem"}}} -->