+++ patterns/header-navigation.php
@@ -9,11 +9,11 @@
 <div class="wp-block-group alignfull">
     <!-- wp:group {"align":"wide","style":{"spacing":{"blockGap":"2em"}},"layout":{"type":"flex","flexWrap":"nowrap","justifyContent":"space-between"}} -->
     <div class="wp-block-group alignwide">
-        <!-- wp:site-title {"level":0,"style":{"typography":{"fontSize":"16px"}}} /-->
+        <!-- wp:site-title {"level":0,"style":{"typography":{"fontSize":"16px","fontWeight":"600","textTransform":"uppercase"}}} /-->
 
         <!-- wp:group {"className":"wp-group","layout":{"type":"flex","flexWrap":"nowrap"}} -->
         <div class="wp-block-group wp-group">
-            <!-- wp:navigation {"ref":8802,"overlayMenu":"never","className":"primary-desktop-navigation","layout":{"type":"flex","justifyContent":"right","orientation":"horizontal","flexWrap":"nowrap"}} /-->
+            <!-- wp:navigation {"ref":8802,"overlayMenu":"never","className":"primary-desktop-navigation","layout":{"type":"flex","justifyContent":"right","orientation":"horizontal","flexWrap":"nowrap"},"style":{"typography":{"fontSize":"14px","fontWeight":"600","textTransform":"uppercase"},"spacing":{"blockGap":"1.75rem"}}} /-->
 
             <!-- wp:buttons {"className":"primary-mobile-navigation"} -->
             <div class="wp-block-buttons primary-mobile-navigation">