+++ patterns/related-posts.php
@@ -5,11 +5,13 @@
  */
 ?>
 
+
 <!-- wp:group {"layout":{"type":"constrained"}} -->
 <div class="wp-block-group">
-    <!-- wp:heading {"textAlign":"center","style":{"typography":{"fontSize":"2.88rem"}},"spacing":{"margin":{"top":"var:preset|spacing|80","bottom":"var:preset|spacing|60"}}}} -->
-    <h2 class="wp-block-heading has-text-align-center" style="margin-top:var(--wp--preset--spacing--80);margin-bottom:var(--wp--preset--spacing--60);font-size:2.88rem"><?php echo __( 'Related posts', 'mdb-theme-fse' ); ?></h2>
+    <!-- wp:heading {"textAlign":"center","style":{"typography":{"fontStyle":"normal","fontWeight":"600"}},"fontSize":"xx-large","spacing":{"margin":{"top":"var(--wp--preset--spacing--80)","bottom":"var(--wp--preset--spacing--60)"}}} -->
+    <h2 class="wp-block-heading has-text-align-center has-xx-large-font-size" style="margin-top:var(--wp--preset--spacing--80);margin-bottom:var(--wp--preset--spacing--60);font-style:normal;font-weight:600;"><?php echo __( 'Related posts', 'mdb-theme-fse' ); ?></h2>
     <!-- /wp:heading -->
+
     <!-- wp:mdb-theme-blocks/related-posts /-->
 </div>
 <!-- /wp:group -->